[API][Admin] Refactor viewing product suite split
