+++ features/product/viewing_product_in_admin_panel/viewing_applied_catalog_promotions_details_for_a_simple_product.feature
@@ -1,4 +1,4 @@
-@viewing_products
+@viewing_product_in_admin_panel
 Feature: Seeing applied catalog promotions details for a simple product
     In order to be aware of simple product price change reason
     As an Administrator
@@ -13,6 +13,7 @@ Feature: Seeing applied catalog promotions details for a simple product
         And it reduces price by "90%"
         And it is enabled
         And I am logged in as an administrator
+        And I am browsing products
 
     @ui @no-api
     Scenario: Seeing applied catalog promotion details on a simple product
