+++ features/product/viewing_product_in_admin_panel/viewing_lowest_price_before_discount_within_variant.feature
@@ -1,4 +1,4 @@
-@viewing_products
+@viewing_product_in_admin_panel
 Feature: Seeing the lowest price before the discount within variant
     In order to be aware of variant's prices
     As an Administrator
