+++ src/Sylius/Behat/Context/Api/Admin/BrowsingCatalogPromotionProductVariantsContext.php
@@ -19,6 +19,9 @@ use Sylius\Behat\Client\ApiClientInterface;
 use Sylius\Behat\Client\ResponseCheckerInterface;
 use Sylius\Behat\Context\Api\Resources;
 use Sylius\Component\Core\Model\CatalogPromotionInterface;
+use Sylius\Component\Core\Model\ChannelInterface;
+use Sylius\Component\Core\Model\ProductInterface;
+use Sylius\Component\Core\Model\ProductVariantInterface;
 use Webmozart\Assert\Assert;
 
 final class BrowsingCatalogPromotionProductVariantsContext implements Context
@@ -37,6 +40,17 @@ final class BrowsingCatalogPromotionProductVariantsContext implements Context
     {
         $this->client->index(Resources::PRODUCT_VARIANTS);
         $this->client->addFilter('catalogPromotion', $this->iriConverter->getIriFromResource($catalogPromotion));
+        $this->client->filter();
+    }
+
+    /**
+     * @When /^I want to view all variants of (this product)$/
+     * @When /^I view(?:| all) variants of the (product "[^"]+")$/
+     */
+    public function iWantToViewAllVariantsOfThisProduct(ProductInterface $product): void
+    {
+        $this->client->index(Resources::PRODUCT_VARIANTS);
+        $this->client->addFilter('product', $this->iriConverter->getIriFromResource($product));
         $this->client->filter();
     }
 
@@ -65,5 +79,72 @@ final class BrowsingCatalogPromotionProductVariantsContext implements Context
                 $variantName,
             ));
         }
+    }
+
+    /**
+     * @Then :variant variant price should be decreased by catalog promotion :catalogPromotion in :channel channel
+     */
+    public function variantPriceShouldBeDecreasedByCatalogPromotion(
+        ProductVariantInterface $variant,
+        CatalogPromotionInterface $catalogPromotion,
+        ChannelInterface $channel,
+    ): void {
+        Assert::true(
+            $this->variantHasCatalogPromotionInChannel($variant, $catalogPromotion, $channel),
+            sprintf(
+                'Catalog promotion "%s" was not found in applied promotions of variant "%s" in channel "%s".',
+                $catalogPromotion->getCode(),
+                $variant->getCode(),
+                $channel->getCode(),
+            ),
+        );
+    }
+
+    /**
+     * @Then :variant variant price should not be decreased by catalog promotion :catalogPromotion in :channel channel
+     */
+    public function variantPriceShouldNotBeDecreasedByCatalogPromotion(
+        ProductVariantInterface $variant,
+        CatalogPromotionInterface $catalogPromotion,
+        ChannelInterface $channel,
+    ): void {
+        Assert::false(
+            $this->variantHasCatalogPromotionInChannel($variant, $catalogPromotion, $channel),
+            sprintf(
+                'Catalog promotion "%s" was found in applied promotions of variant "%s" in channel "%s".',
+                $catalogPromotion->getCode(),
+                $variant->getCode(),
+                $channel->getCode(),
+            ),
+        );
+    }
+
+    private function variantHasCatalogPromotionInChannel(
+        ProductVariantInterface $variant,
+        CatalogPromotionInterface $catalogPromotion,
+        ChannelInterface $channel
+    ): bool {
+        $variantData = $this->getDataOfVariantWithCode($variant->getCode());
+
+        $promotions = $variantData['channelPricings'][$channel->getCode()]['appliedPromotions'] ?? [];
+        foreach ($promotions as $promotion) {
+            if ($promotion['code'] === $catalogPromotion->getCode()) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    private function getDataOfVariantWithCode(string $code): array
+    {
+        $variantsData = $this->responseChecker->getCollection($this->client->getLastResponse());
+        foreach ($variantsData as $variantData) {
+            if ($variantData['code'] === $code) {
+                return $variantData;
+            }
+        }
+
+        throw new \InvalidArgumentException(sprintf('Variant with code "%s" was not found.', $code));
     }
 }
