+++ src/Sylius/Behat/Context/Api/Admin/ManagingProductVariantsContext.php
@@ -733,71 +733,4 @@ final class ManagingProductVariantsContext implements Context
             'On hand must be greater than the number of on hold units',
         );
     }
-
-    /**
-     * @Then :variant variant price should be decreased by catalog promotion :catalogPromotion in :channel channel
-     */
-    public function variantPriceShouldBeDecreasedByCatalogPromotion(
-        ProductVariantInterface $variant,
-        CatalogPromotionInterface $catalogPromotion,
-        ChannelInterface $channel,
-    ): void {
-        Assert::true(
-            $this->variantHasCatalogPromotionInChannel($variant, $catalogPromotion, $channel),
-            sprintf(
-                'Catalog promotion "%s" was not found in applied promotions of variant "%s" in channel "%s".',
-                $catalogPromotion->getCode(),
-                $variant->getCode(),
-                $channel->getCode(),
-            ),
-        );
-    }
-
-    /**
-     * @Then :variant variant price should not be decreased by catalog promotion :catalogPromotion in :channel channel
-     */
-    public function variantPriceShouldNotBeDecreasedByCatalogPromotion(
-        ProductVariantInterface $variant,
-        CatalogPromotionInterface $catalogPromotion,
-        ChannelInterface $channel,
-    ): void {
-        Assert::false(
-            $this->variantHasCatalogPromotionInChannel($variant, $catalogPromotion, $channel),
-            sprintf(
-                'Catalog promotion "%s" was found in applied promotions of variant "%s" in channel "%s".',
-                $catalogPromotion->getCode(),
-                $variant->getCode(),
-                $channel->getCode(),
-            ),
-        );
-    }
-
-    private function variantHasCatalogPromotionInChannel(
-        ProductVariantInterface $variant,
-        CatalogPromotionInterface $catalogPromotion,
-        ChannelInterface $channel
-    ): bool {
-        $variantData = $this->getDataOfVariantWithCode($variant->getCode());
-
-        $promotions = $variantData['channelPricings'][$channel->getCode()]['appliedPromotions'] ?? [];
-        foreach ($promotions as $promotion) {
-            if ($promotion['code'] === $catalogPromotion->getCode()) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    private function getDataOfVariantWithCode(string $code): array
-    {
-        $variantsData = $this->responseChecker->getCollection($this->client->getLastResponse());
-        foreach ($variantsData as $variantData) {
-            if ($variantData['code'] === $code) {
-                return $variantData;
-            }
-        }
-
-        throw new \InvalidArgumentException(sprintf('Variant with code "%s" was not found.', $code));
-    }
 }
