+++ src/Sylius/Behat/Resources/config/suites/api/product/viewing_product_in_admin_panel.yaml
@@ -8,20 +8,38 @@ default:
                 - sylius.behat.context.hook.doctrine_orm
 
                 - sylius.behat.context.transform.channel
+                - sylius.behat.context.transform.currency
                 - sylius.behat.context.transform.lexical
+                - sylius.behat.context.transform.locale
                 - sylius.behat.context.transform.product
+                - sylius.behat.context.transform.product_association_type
+                - sylius.behat.context.transform.product_attribute
                 - sylius.behat.context.transform.product_option
+                - sylius.behat.context.transform.product_option_value
+                - sylius.behat.context.transform.product_variant
+                - sylius.behat.context.transform.shared_storage
+                - sylius.behat.context.transform.shipping_category
+                - sylius.behat.context.transform.tax_category
                 - sylius.behat.context.transform.taxon
+                - Sylius\Behat\Context\Transform\CatalogPromotionContext
 
                 - sylius.behat.context.setup.admin_api_security
                 - sylius.behat.context.setup.channel
+                - sylius.behat.context.setup.locale
                 - sylius.behat.context.setup.product
                 - sylius.behat.context.setup.product_association
+                - sylius.behat.context.setup.product_attribute
                 - sylius.behat.context.setup.product_option
+                - sylius.behat.context.setup.product_taxon
                 - sylius.behat.context.setup.shipping_category
+                - sylius.behat.context.setup.taxation
                 - sylius.behat.context.setup.taxonomy
+                - Sylius\Behat\Context\Setup\CatalogPromotionContext
+                - Sylius\Behat\Context\Setup\PriceHistoryContext
+                - Sylius\Calendar\Tests\Behat\Context\Setup\CalendarContext
 
                 - sylius.behat.context.api.admin.managing_products
+                - Sylius\Behat\Context\Api\Admin\BrowsingCatalogPromotionProductVariantsContext
 
             filters:
                 tags: "@viewing_product_in_admin_panel&&@api"
