+++ src/Sylius/Behat/Resources/config/suites/ui/product/viewing_product_in_admin_panel.yaml
@@ -9,7 +9,9 @@ default:
                 - sylius.behat.context.hook.session
 
                 - sylius.behat.context.transform.channel
+                - sylius.behat.context.transform.currency
                 - sylius.behat.context.transform.lexical
+                - sylius.behat.context.transform.locale
                 - sylius.behat.context.transform.product
                 - sylius.behat.context.transform.product_association_type
                 - sylius.behat.context.transform.product_option
@@ -19,19 +21,26 @@ default:
                 - sylius.behat.context.transform.shipping_category
                 - sylius.behat.context.transform.tax_category
                 - sylius.behat.context.transform.taxon
+                - Sylius\Behat\Context\Transform\CatalogPromotionContext
 
                 - sylius.behat.context.setup.admin_security
                 - sylius.behat.context.setup.channel
+                - sylius.behat.context.setup.locale
                 - sylius.behat.context.setup.product
                 - sylius.behat.context.setup.product_association
+                - sylius.behat.context.setup.product_attribute
                 - sylius.behat.context.setup.product_option
                 - sylius.behat.context.setup.product_taxon
                 - sylius.behat.context.setup.shipping_category
-                - sylius.behat.context.setup.taxonomy
                 - sylius.behat.context.setup.taxation
+                - sylius.behat.context.setup.taxonomy
+                - Sylius\Behat\Context\Setup\CatalogPromotionContext
+                - Sylius\Behat\Context\Setup\PriceHistoryContext
+                - Sylius\Calendar\Tests\Behat\Context\Setup\CalendarContext
 
                 - sylius.behat.context.ui.admin.managing_product_attributes
                 - sylius.behat.context.ui.admin.product_showpage
+                - sylius.behat.context.ui.shop.browsing_product
 
             filters:
                 tags: "@viewing_product_in_admin_panel&&@ui"
